requirements.txt

Django>=2.1.0,<2.2.0
djangorestframework>=3.8.2,<3.9.0


flake8>=3.6.0,<3.7.0

